% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_model}
\alias{predict_model}
\title{A prediction Function}
\usage{
predict_model(data, model)
}
\arguments{
\item{data}{a matrix of processed integer sequences corresponding to cheif complaints}

\item{model}{A keras Model (hdf5)}
}
\description{
This function makes class predictions based on the prepocessed data and an uploaded model
}
\examples{
predict_model()
}
\keyword{predict_model}
