% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_thresh.R
\name{predict_thresh}
\alias{predict_thresh}
\title{A prediction function based on a predetermined decision threshold}
\usage{
predict_thresh(data, model)
}
\arguments{
\item{data}{a matrix of processed integer sequences corresponding to chief complaints}

\item{model}{A keras Model (hdf5)}
}
\description{
This function makes class predictions based on the prepocessed data and an uploaded model
with a predetermined threshold
}
\keyword{threshold}
