% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{A data pre-processing Function}
\usage{
process(x)
}
\arguments{
\item{x}{a dataframe of strings}
}
\description{
This function processes chief complaints and generates a sequence of integers based on the
}
\examples{
process()
}
\keyword{process}
